Implement projectile damage system with collision tracking

+++ src/types/health.odin
@@ -7,6 +7,6 @@ DEFAULT_CURRENT_HEALTH :: DEFAULT_MAX_HEALTH
 DEFAULT_HEALTH: Health : {max = DEFAULT_MAX_HEALTH, current = DEFAULT_CURRENT_HEALTH}
 
 Health :: struct {
-	max:     u32,
-	current: u32,
+	max:     f32,
+	current: f32,
 }
